# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: FivePeaks Payment Engine API
  version: 2.0.0
  description: >-
    Tenant-scoped API for Mercado Pago OAuth, Brick payments, order status,
    and incoming Mercado Pago payment notifications.
servers:
  - url: https://{host}
    variables:
      host:
        default: payments.example.com
security:
  - TenantApiKey: []
paths:
...
  /v2/ordenes/{id}/estado:
    get:
      operationId: getOrderStatus
      summary: Get an order by ID within the authenticated tenant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Current order and latest payment attempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Order does not exist for this tenant. }
...
components:
  securitySchemes:
    TenantApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    OrderStatus:
      type: object
      required: [id_orden, external_id, estado, mp_payment_id, updated_at, ultimo_pago]
      properties:
        id_orden: { type: string, format: uuid }
        external_id: { type: string }
        estado: { $ref: '#/components/schemas/PaymentStatus' }
        mp_payment_id: { type: [string, 'null'] }
        updated_at: { type: string, format: date-time }
        ultimo_pago:
          oneOf:
            - type: 'null'
            - type: object
              required: [id_pago, estado, status_detail, mp_payment_id]
              properties:
                id_pago: { type: string, format: uuid }
                estado: { $ref: '#/components/schemas/PaymentStatus' }
                status_detail: { type: [string, 'null'] }
                mp_payment_id: { type: [string, 'null'] }
    PaymentStatus:
      type: string
      enum: [pending, approved, rejected, cancelled, refunded, charged_back]
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
  responses:
...
    Unauthorized:
      description: API key is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...